temp: Modify Questionnaires Index Type to `int`

+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/questionnaire/GetAssessmentNextQuestionnaireResponseDto.java
@@ -31,7 +31,7 @@ public record GetAssessmentNextQuestionnaireResponseDto(ResultStatus status, Dat
         }
 
         @Builder
-        record Found(long id, long index, String title) implements Data {
+        record Found(long id, int index, String title) implements Data {
 
             public static Found of(GetAssessmentNextQuestionnaireUseCase.Result.Found data) {
                 return Found.builder()
